restore to / from disk / ram, either way works
also move cached disk to separate file, other changes
